requirements.txt
asgiref==3.9.1
certifi==2025.7.9
charset-normalizer==3.4.2
cloudinary==1.44.1
colorama==0.4.6
coverage==7.9.2
Django==5.2.4
django-cloudinary-storage==0.3.0
django-filter==25.1
django-widget-tweaks==1.5.0
Faker==37.5.3
gunicorn==23.0.0
idna==3.10
iniconfig==2.1.0
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
Pygments==2.19.2
pytest==8.4.1
pytest-cov==6.2.1
pytest-django==4.11.1
python-dotenv==1.1.1
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.9.0
